requirements.txt
requests==2.25.1
psycopg2-binary==2.8.6
python-dotenv==0.15.0
